fix:[result] 권한별 라우팅 설정 add:[result]GuestPage파일 생성
