+++ next.config.js
@@ -2,7 +2,11 @@
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  /* config options here */
+  images: {
+    domains: [
+      'sfdcyhvieqruoagzezzv.supabase.co', // Add this line
+    ],
+  },
 };
 
 module.exports = nextConfig;
